+++ RELEASE_NOTES.md
@@ -1,3 +1,11 @@
+### Version 1.7.0
+__Changes__
+- updated BLAST to 2.13.0
+- fixed bug with long feature IDs
+- added SpeciesTree as possible target
+- add option to write off-genetic-code protein translations
+- tightened up client instantiation
+
 ### Version 1.6.0
 __Changes__
 - updated BLAST to 2.12.0